feat: deleted unused APIs and moved some lib / APIs to different folders
