+++ pages/api/account/register.ts
